Move navbar components
